Fix ArbitraryIntrospector name
