+++ fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/introspector/ArbitraryTypeIntrospector.java
@@ -18,23 +18,71 @@
 
 package com.navercorp.fixturemonkey.api.introspector;
 
-import java.util.Arrays;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import org.apiguardian.api.API;
 import org.apiguardian.api.API.Status;
 
+import net.jqwik.api.Arbitrary;
+import net.jqwik.api.arbitraries.BigDecimalArbitrary;
+import net.jqwik.api.arbitraries.BigIntegerArbitrary;
+import net.jqwik.api.arbitraries.ByteArbitrary;
+import net.jqwik.api.arbitraries.CharacterArbitrary;
+import net.jqwik.api.arbitraries.DoubleArbitrary;
+import net.jqwik.api.arbitraries.FloatArbitrary;
+import net.jqwik.api.arbitraries.IntegerArbitrary;
+import net.jqwik.api.arbitraries.LongArbitrary;
+import net.jqwik.api.arbitraries.ShortArbitrary;
+import net.jqwik.api.arbitraries.StringArbitrary;
+
 import com.navercorp.fixturemonkey.api.generator.ArbitraryGeneratorContext;
 
 @API(since = "0.4.0", status = Status.EXPERIMENTAL)
-@FunctionalInterface
 public interface ArbitraryTypeIntrospector {
-	ArbitraryTypeIntrospector INTROSPECTORS = new CompositeArbitraryTypeIntrospector(
-		Arrays.asList(
-			new EnumTypeIntrospector(),
-			new BooleanTypeIntrospector(),
-			new UuidTypeIntrospector()
-		)
-	);
+	default Arbitrary<String> strings(StringArbitrary stringArbitrary, ArbitraryGeneratorContext context) {
+		return stringArbitrary;
+	}
 
-	ArbitraryIntrospectorResult introspect(ArbitraryGeneratorContext context);
+	default Arbitrary<Character> characters(CharacterArbitrary characterArbitrary, ArbitraryGeneratorContext context) {
+		return characterArbitrary;
+	}
+
+	default Arbitrary<Short> shorts(ShortArbitrary shortArbitrary, ArbitraryGeneratorContext context) {
+		return shortArbitrary;
+	}
+
+	default Arbitrary<Byte> bytes(ByteArbitrary byteArbitrary, ArbitraryGeneratorContext context) {
+		return byteArbitrary;
+	}
+
+	default Arbitrary<Double> doubles(DoubleArbitrary doubleArbitrary, ArbitraryGeneratorContext context) {
+		return doubleArbitrary;
+	}
+
+	default Arbitrary<Float> floats(FloatArbitrary floatArbitrary, ArbitraryGeneratorContext context) {
+		return floatArbitrary;
+	}
+
+	default Arbitrary<Integer> integers(IntegerArbitrary integerArbitrary, ArbitraryGeneratorContext context) {
+		return integerArbitrary;
+	}
+
+	default Arbitrary<Long> longs(LongArbitrary longArbitrary, ArbitraryGeneratorContext context) {
+		return longArbitrary;
+	}
+
+	default Arbitrary<BigInteger> bigIntegers(
+		BigIntegerArbitrary bigIntegerArbitrary,
+		ArbitraryGeneratorContext context
+	) {
+		return bigIntegerArbitrary;
+	}
+
+	default Arbitrary<BigDecimal> bigDecimals(
+		BigDecimalArbitrary bigDecimalArbitrary,
+		ArbitraryGeneratorContext context
+	) {
+		return bigDecimalArbitrary;
+	}
 }
